상품의 후기 조회 진행 중 #217

+++ src/main/java/com/danakga/webservice/review/dto/request/ReqReviewDto.java
@@ -28,6 +28,6 @@ public class ReqReviewDto {
     private String reviewContent;
 
     //후기 별점
-    private Double reviewScore;
+    private int reviewScore;
 
 }
